% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\name{.add_gg}
\alias{.add_gg}
\title{Override + operator for ggplot objects to log calls}
\usage{
.add_gg(e1, e2)
}
\arguments{
\item{e1}{A ggplot object.}

\item{e2}{A layer to add to the ggplot object.}
}
\value{
A ggplot object with logged calls.
}
\description{
Override + operator for ggplot objects to log calls
}
\keyword{internal}
