% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\docType{class}
\name{GGLog-class}
\alias{GGLog-class}
\alias{GGLog}
\title{GGLog class}
\description{
The `GGLog` class is used to store a single ggplot call.
}
\section{Fields}{

\describe{
\item{\code{code}}{A string to store the code of the ggplot call.}
}}

\section{Methods}{

\describe{
\item{\code{evaluate(envir = parent.frame())}}{Evaluate the log.

      @param envir The environment to evaluate the log in.}

\item{\code{stringify()}}{Stringify the log.

      @return A string.}
}}

