% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\docType{class}
\name{GGLogs-class}
\alias{GGLogs-class}
\alias{GGLogs}
\title{GGLogs class}
\description{
The `GGLogs` class is used to store logs of ggplot calls.
  Each log is stored as a `GGLog` object.
  The object can be accessed using the `$logs` field from a ggplot object.
}
\section{Fields}{

\describe{
\item{\code{logs}}{A list to store the logs.}
}}

\section{Methods}{

\describe{
\item{\code{add(log)}}{Add a log to the list.

        @param log A GGLog object.}

\item{\code{evaluate(env = parent.frame())}}{Evaluate all logs in the list.

      @param env The environment to evaluate the logs in.}

\item{\code{gen_code(setup = "library(ggplot2)")}}{Generate code for all logs in the list.

      @param setup A string to setup the environment.

      @return A string of code.}
}}

