% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\name{ggplot}
\alias{ggplot}
\title{Override ggplot function to log calls}
\usage{
ggplot(...)
}
\arguments{
\item{...}{Arguments passed to ggplot2::ggplot.}
}
\value{
A ggplot object with logged calls.
}
\description{
Override ggplot function to log calls
}
