% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\name{print.GGLogs}
\alias{print.GGLogs}
\title{Print a GGLogs object}
\usage{
\method{print}{GGLogs}(x, ...)
}
\arguments{
\item{x}{A GGLogs object.}

\item{...}{Additional arguments.}
}
\value{
Display the logs and invisibly return the logs.
}
\description{
Print a GGLogs object
}
