% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\name{register}
\alias{register}
\title{Register a function that returns a ggplot object
This is usually implemented by ggplot extensions.}
\usage{
register(f, name = NULL)
}
\arguments{
\item{f}{A function that returns a ggplot object.}
}
\value{
A function that returns a ggplot object with logged calls.
}
\description{
Register a function that returns a ggplot object
This is usually implemented by ggplot extensions.
}
