% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\docType{class}
\name{GGLog-class}
\alias{GGLog-class}
\alias{GGLog}
\title{GGLog class}
\description{
The `GGLog` class is used to store a single ggplot call.
}
\section{Fields}{

\describe{
\item{\code{code}}{A string to store the code of the ggplot call.}
}}

\section{Methods}{

\describe{
\item{\code{evaluate(env = parent.frame())}}{Evaluate the log.

      @param env The environment to evaluate the log in.}
}}

