% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglogger.R
\name{print.GGLog}
\alias{print.GGLog}
\title{Print a GGLog object}
\usage{
\method{print}{GGLog}(x, ...)
}
\arguments{
\item{x}{A GGLog object.}

\item{...}{Additional arguments.}
}
\description{
Print a GGLog object
}
